+++ scripts/generate-schema.ts
@@ -4,6 +4,7 @@ import { match, P } from "npm:ts-pattern";
 import type {
   JSONSchema7 as JSONSchema,
   JSONSchema7Definition as JSONSchemaDefinition,
+  JSONSchema7TypeName,
 } from "npm:@types/json-schema";
 
 const schemasDir = new URL("../schemas", import.meta.url).pathname;
@@ -36,6 +37,8 @@ type NodeIR =
   | { type: "boolean"; optional?: boolean }
   | { type: "string"; optional?: boolean }
   | { type: "literal"; value: string }
+  | { type: "int"; minimum?: number; maximum?: number }
+  | { type: "float"; minimum?: number; maximum?: number }
   | { type: "unknown" };
 
 const isDescriminatedUnion = (def: JSONSchemaDefinition[] | undefined) => {
@@ -43,20 +46,58 @@ const isDescriminatedUnion = (def: JSONSchemaDefinition[] | undefined) => {
     (def[0]?.required?.[0] + "").startsWith("$");
 };
 
+const isOptionalType =
+  (typeOf: string) =>
+  (type: JSONSchema7TypeName | JSONSchema7TypeName[] | undefined) => {
+    if (type && Array.isArray(type) && type[0] === typeOf) {
+      return true;
+    }
+    return false;
+  };
+
 function jsonSchemaToIR(schema: JSONSchema): DocIR {
   const nodeToIR = (node: JSONSchema): NodeIR => {
     return match(node)
       .with(
-        { type: "boolean" },
+        {
+          type: P.union("boolean", P.when(isOptionalType("boolean"))),
+        },
         (node) =>
-          ({ type: "boolean" as const, optional: !!node.default }) as const,
+          ({
+            type: "boolean" as const,
+            optional: !!node.default,
+          }) as const,
       )
+      .with({ type: "integer" }, (node) => ({
+        type: "int" as const,
+        minimum: node.minimum,
+        maximum: node.maximum,
+      }))
+      .with({ type: "number", format: "double" }, (node) => ({
+        type: "float" as const,
+        minimum: node.minimum,
+        maximum: node.maximum,
+      }))
       .with(
         { type: "string" },
-        (node) =>
-          node.enum
-            ? ({ type: "literal" as const, value: node.enum[0] as string })
-            : ({ type: "string" as const, optional: !!node.default }),
+        (node) => {
+          if (node.enum) {
+            if (node.enum.length === 1) {
+              return {
+                type: "literal" as const,
+                value: node.enum[0] as string,
+              };
+            }
+            return {
+              type: "union" as const,
+              members: node.enum.map((v) => ({
+                type: "literal" as const,
+                value: v as string,
+              })),
+            };
+          }
+          return ({ type: "string" as const, optional: !!node.default });
+        },
       )
       .with(
         { oneOf: P.when(isDescriminatedUnion) },
@@ -76,7 +117,7 @@ function jsonSchemaToIR(schema: JSONSchema): DocIR {
       .with(
         { oneOf: P.array() },
         (node) => {
-          const intersection = {
+          const union = {
             type: "union" as const,
             members: node.oneOf?.map((v) => nodeToIR(v as JSONSchema)) ?? [],
           };
@@ -95,18 +136,26 @@ function jsonSchemaToIR(schema: JSONSchema): DocIR {
                     value: nodeToIR(value as JSONSchema),
                   })),
                 },
-                intersection,
+                union,
               ],
             });
           }
-          return intersection;
+          return union;
         },
       )
       .with(
         { anyOf: P.array() },
         () => ({
           type: "union" as const,
-          members: (node.anyOf?.map((v) => nodeToIR(v as JSONSchema)) ?? []),
+          members: (node.anyOf?.map((v) => nodeToIR(v as JSONSchema)) ?? [])
+            .filter((v) => v.type !== "unknown")
+            // flatten nested unions
+            .reduce((union, member) => {
+              if (member.type === "union") {
+                return union.concat(member.members);
+              }
+              return union.concat(member);
+            }, [] as NodeIR[]),
         }),
       )
       .with(
@@ -115,11 +164,13 @@ function jsonSchemaToIR(schema: JSONSchema): DocIR {
           type: "object" as const,
           properties: Object.entries(node.properties ?? {}).map((
             [key, value],
-          ) => ({
-            key,
-            required: node.required?.includes(key) ?? false,
-            value: nodeToIR(value as JSONSchema),
-          })),
+          ) => {
+            return ({
+              key,
+              required: node.required?.includes(key) ?? false,
+              value: nodeToIR(value as JSONSchema),
+            });
+          }),
         }),
       )
       .otherwise(() => ({ type: "unknown" }));
@@ -147,6 +198,8 @@ function generateTypes(ir: DocIR) {
 
   function generateNode(node: NodeIR) {
     match(node)
+      .with({ type: "int" }, () => w("number"))
+      .with({ type: "float" }, () => w("number"))
       .with({ type: "boolean" }, () => w("boolean"))
       .with({ type: "string" }, () => w("string"))
       .with({ type: "literal" }, (node) => w(`"${node.value}"`))
@@ -205,6 +258,24 @@ function generateZodSchema(ir: DocIR) {
 
   function generateNode(node: NodeIR) {
     match(node)
+      .with({ type: "int" }, (node) => {
+        w("z.number().int()");
+        if (typeof node.minimum === "number") {
+          w(`.min(${node.minimum})`);
+        }
+        if (typeof node.maximum === "number") {
+          w(`.max(${node.maximum})`);
+        }
+      })
+      .with({ type: "float" }, (node) => {
+        w("z.number()");
+        if (typeof node.minimum === "number") {
+          w(`.min(${node.minimum})`);
+        }
+        if (typeof node.maximum === "number") {
+          w(`.max(${node.maximum})`);
+        }
+      })
       .with(
         { type: "boolean" },
         (node) => w("z.boolean()", node.optional && ".optional()"),